+++ common/infallible/tests/it/main.rs
@@ -12,29 +12,5 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#[test]
-fn test_rwlock() {
-    use std::sync::Arc;
-    use std::thread;
-
-    use crate::RwLock;
-
-    let a = 7u8;
-    let rwlock = Arc::new(RwLock::new(a));
-    let rwlock2 = rwlock.clone();
-    let rwlock3 = rwlock.clone();
-
-    let thread1 = thread::spawn(move || {
-        let mut b = rwlock2.write();
-        *b = 8;
-    });
-    let thread2 = thread::spawn(move || {
-        let mut b = rwlock3.write();
-        *b = 9;
-    });
-
-    let _ = thread1.join();
-    let _ = thread2.join();
-
-    let _read = rwlock.read();
-}
+mod mutex;
+mod rwlock;
